requirements.txt
# Temporal SDK
temporalio>=1.5.0

# MCP Server
fastmcp>=0.1.0

# LLM Integration (Groq - Free & Fast)
groq>=0.4.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0